Avoid stack too deep error

+++ contracts/interfaces/IvPair.sol
@@ -69,7 +69,7 @@ interface IvPair {
         uint256 amountOut,
         address ikPair,
         address to,
-        uint8 incentivesLimitPct,
+        uint256 incentivesLimitPct,
         bytes calldata data
     ) external returns (address _token, uint256 _leftovers);
